Fix statusline first buffer alignment
When using statusline integration, previously the first buffer had one
space to its left when it was current and none when it was not. This
made the alignment change for all buffers when switching to and from the
first buffer.

This change fixes that by making it so that the first buffer always has
one space to the left.